More precise info about Data Insertion API
Bot this doc and the linked Data Insertion API docs miss which parameter shall be used in the Data Insertion API. As far as Adobe support said, it is <a4t>. Also this document simply speaks about the "payload", but if I understand Adobe support correctly, it is precisely the payload.tnta parameter.

Furthermore, in my test, the property in the Response tokens is called `analyticsDetails`, not just `analytics`. And it is an Array. 

All this should be verified by a Target expert at Adobe before accepting the PR.